Class extraction for API. Added getChannels from project method
