+++ src/main/java/com/octopusdeploy/api/EnvironmentsApi.java
@@ -1,0 +1,72 @@
+package com.octopusdeploy.api;
+
+import com.octopusdeploy.api.data.Environment;
+import java.io.IOException;
+import java.util.HashSet;
+import java.util.Set;
+import net.sf.json.JSONArray;
+import net.sf.json.JSONObject;
+import net.sf.json.JSONSerializer;
+
+public class EnvironmentsApi {
+    private final static String UTF8 = "UTF-8";
+    private final AuthenticatedWebClient webClient;
+
+    public EnvironmentsApi(AuthenticatedWebClient webClient) {
+        this.webClient = webClient;
+    }
+
+    /**
+     * Get all environments from the Octopus server as Environment objects.
+     * @return A set of all environments on the Octopus server.
+     * @throws IllegalArgumentException when the web client receives a bad parameter
+     * @throws IOException When the AuthenticatedWebClient receives and error response code
+     */
+    public Set<Environment> getAllEnvironments() throws IllegalArgumentException, IOException {
+        HashSet<Environment> environments = new HashSet<Environment>();
+        AuthenticatedWebClient.WebResponse response =webClient.get("api/environments/all");
+        if (response.isErrorCode()) {
+            throw new IOException(String.format("Code %s - %n%s", response.getCode(), response.getContent()));
+        }
+        JSONArray json = (JSONArray)JSONSerializer.toJSON(response.getContent());
+        for (Object obj : json) {
+            JSONObject jsonObj = (JSONObject)obj;
+            String id = jsonObj.getString("Id");
+            String name = jsonObj.getString("Name");
+            String description = jsonObj.getString("Description");
+            environments.add(new Environment(id, name, description));
+        }
+        return environments;
+    }
+
+    /**
+     * Get the Environment with the given name if it exists, return null otherwise.
+     * Only selects the environment if the name is an exact match (including case)
+     * @param name The name of the Environment to find.
+     * @return The Environment with that name.
+     * @throws IllegalArgumentException when the web client receives a bad parameter
+     * @throws IOException When the AuthenticatedWebClient receives and error response code
+     */
+    public Environment getEnvironmentByName(String name) throws IllegalArgumentException, IOException {
+        return getEnvironmentByName(name, false);
+    }
+
+    /**
+     * Get the Environment with the given name if it exists, return null otherwise.
+     * @param name The name of the Environment to find.
+     * @param ignoreCase when true uses equalsIgnoreCase in the name check
+     * @return The Environment with that name.
+     * @throws IllegalArgumentException when the web client receives a bad parameter
+     * @throws IOException When the AuthenticatedWebClient receives and error response code
+     */
+    public Environment getEnvironmentByName(String name, boolean ignoreCase) throws IllegalArgumentException, IOException {
+        Set<Environment> environments = getAllEnvironments();
+        for (Environment env : environments) {
+            if ((ignoreCase && name.equalsIgnoreCase(env.getName())) ||
+               (!ignoreCase && name.equals(env.getName()))) {
+                return env;
+            }
+        }
+        return null;
+    }
+}
