+++ src/main/java/com/octopusdeploy/api/ProjectsApi.java
@@ -1,0 +1,71 @@
+package com.octopusdeploy.api;
+
+import com.octopusdeploy.api.data.Project;
+import java.io.IOException;
+import java.util.HashSet;
+import java.util.Set;
+import net.sf.json.JSONArray;
+import net.sf.json.JSONObject;
+import net.sf.json.JSONSerializer;
+
+public class ProjectsApi {
+    private final AuthenticatedWebClient webClient;
+
+    public ProjectsApi(AuthenticatedWebClient webClient) {
+        this.webClient = webClient;
+    }
+
+    /**
+     * Uses the authenticated web client to pull all projects from the api and
+     * convert them to POJOs
+     * @return a Set of Projects (may be empty)
+     * @throws IllegalArgumentException when the web client receives a bad parameter
+     * @throws IOException When the AuthenticatedWebClient receives and error response code
+     */
+    public Set<Project> getAllProjects() throws IllegalArgumentException, IOException {
+        HashSet<Project> projects = new HashSet<Project>();
+        AuthenticatedWebClient.WebResponse response = webClient.get("api/projects/all");
+        if (response.isErrorCode()) {
+            throw new IOException(String.format("Code %s - %n%s", response.getCode(), response.getContent()));
+        }
+        JSONArray json = (JSONArray)JSONSerializer.toJSON(response.getContent());
+        for (Object obj : json) {
+            JSONObject jsonObj = (JSONObject)obj;
+            String id = jsonObj.getString("Id");
+            String name = jsonObj.getString("Name");
+            projects.add(new Project(id, name));
+        }
+        return projects;
+    }
+
+    /**
+     * Loads in the full list of projects from the API, then selects one project by name.
+     * Only selects the project if the name is an exact match (including case)
+     * @param name name of the project to select
+     * @return the named project or null if no such project exists
+     * @throws IllegalArgumentException when the web client receives a bad parameter
+     * @throws IOException When the AuthenticatedWebClient receives and error response code
+     */
+    public Project getProjectByName(String name)  throws IllegalArgumentException, IOException {
+        return getProjectByName(name, false);
+    }
+
+    /**
+     * Loads in the full list of projects from the API, then selects one project by name.
+     * @param name name of the project to select
+     * @param ignoreCase when true uses equalsIgnoreCase in the name check
+     * @return the named project or null if no such project exists
+     * @throws IllegalArgumentException when the web client receives a bad parameter
+     * @throws IOException When the AuthenticatedWebClient receives and error response code
+     */
+    public Project getProjectByName(String name, boolean ignoreCase)  throws IllegalArgumentException, IOException {
+        Set<Project> allProjects = getAllProjects();
+        for (Project project : allProjects) {
+            if ((ignoreCase && name.equalsIgnoreCase(project.getName())) ||
+               (!ignoreCase && name.equals(project.getName()))) {
+                return project;
+            }
+        }
+        return null;
+    }
+}
